// src/api/generated/models/createActivityBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Campus Platform API
 * OpenAPI spec version: 1.1.0
 */

export type CreateActivityBody = {
  title: string;
  summary: string;
  body: string;
  location: string;
  signup_start_at: string;
  signup_end_at: string;
  start_at: string;
  end_at: string;
  /** @minimum 1 */
  capacity: number;
  contact_type: string;
  contact: string;
};