// src/api/generated/models/createCampusCirclePostBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Campus Platform API
 * OpenAPI spec version: 1.1.0
 */

export type CreateCampusCirclePostBody = {
  /** @minimum 1 */
  section_id: number;
  /** @maxLength 100 */
  title?: string;
  /** @maxLength 5000 */
  content?: string;
  image_urls?: string[];
};