// src/api/generated/models/createCarpoolTripBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Campus Platform API
 * OpenAPI spec version: 1.1.0
 */

export type CreateCarpoolTripBody = {
  /** @maxLength 200 */
  title: string;
  /** @maxLength 500 */
  origin: string;
  /** @maxLength 500 */
  destination: string;
  departure_at: string;
  /** @minimum 1 */
  total_seats: number;
  /** @maxLength 16 */
  contact_type: string;
  /** @maxLength 200 */
  contact: string;
};