// src/api/generated/models/createErrandBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Campus Platform API
 * OpenAPI spec version: 1.1.0
 */

export type CreateErrandBody = {
  title: string;
  description: string;
  /** @minimum 1 */
  reward_cents: number;
  pickup_location: string;
  dropoff_location: string;
  deadline: string;
  contact_type: string;
  contact: string;
};