// src/api/generated/models/createMarketplaceListingBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Campus Platform API
 * OpenAPI spec version: 1.1.0
 */

export type CreateMarketplaceListingBody = {
  title: string;
  description: string;
  /** @minimum 1 */
  price_cents: number;
  contact_type: string;
  contact: string;
  image_urls?: string[];
};