// src/api/generated/models/campusCircleSectionView.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Campus Platform API
 * OpenAPI spec version: 1.1.0
 */
import type { CampusCircleSectionStatus } from './campusCircleSectionStatus';

export interface CampusCircleSectionView {
  children: CampusCircleSectionView[];
  /** @nullable */
  cover_url: string | null;
  created_at: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  icon_url: string | null;
  id: number;
  name: string;
  /** @nullable */
  parent_id: number | null;
  slug: string;
  sort_order: number;
  status: CampusCircleSectionStatus;
  updated_at: string;
  version: number;
}